Generalize `seqio.Evaluator` logging to support multiple message types.
`metric_fn`s that are attached to various t5 tasks used to return
`Mapping[str, float]` where the value represents a new datapoint to be added to
tensorboard and the key is the name of the series that the new datapoint will
be added to. This updates the logging facilities so that `metric_fn`s can
instead return a `Mapping[str, Metric]`, where the `Metric` is a baseclass of
several data types supported by Tensorboard. Then custom logging functions can
dispatch to appropreate implemenation of writing that data type to tensorboard.

When a metric function hasn't been updated to ouput `Metric`s instead of
`float`s the Evaluator will wrap the returned floats in `Scalar` objects.

PiperOrigin-RevId: 353045211