Add custom logging to `seqio.evaluation.Evaluator`.
This CL adds the ability to pass a function to the Evaluator via the `log_fn`
parameter. When running evaluation this function with the resulting task
metrics. If a `log_fn` isn't provided it will default to using the
`TensorboardLogging` object in `seqio.evaluation` which serves as example of a
custom logging. This object has equivalent functionality to the old
`_log_eval_results` method.

This also includes some updates the the evaluation tests so the default
`_log_fn` is set when we mock the `Evaluator`'s `__init__` As well as tests for
the cases where both a `summary_dir` and a `log_fn` are provided to the
`Evaluator`

PiperOrigin-RevId: 353045215